fix(api): type NodeStat mode/nlink as uint32 to match POSIX
Before this commit:
- The Security Scan CI job (gosec) failed on PR #23 with two
  G115 (CWE-190) warnings: int32(n.Mode()) and int32(n.NLink())
  in handler/fs.go's statToAPI helper. gosec flags every
  uint32 → int32 conversion as a potential integer-overflow
  even when the source value is bounded.
- The block was on the chain: Security Scan fails, Unit /
  Integration / E2E / Docker Build all skip, the PR cannot
  merge under the 7/7-green policy.

This commit:
- Changes the NodeStat schema's mode and nlink fields from
  int32 to uint32 in api/schemas/node.yaml. POSIX st_mode and
  st_nlink are unsigned 32-bit values; the previous int32
  was a glibc bias (st_nlink is signed in glibc, unsigned in
  POSIX) that the audit and the security linter both reject.
- The handler now assigns the uint32 field directly, so the
  cast that gosec was flagging no longer happens.
- Regenerates pkg/api from the bundled OpenAPI. ogen maps
  the 'format: uint32' extension to a Go uint32 field.

+++ internal/app/apiserver/handler/fs.go
@@ -278,8 +278,8 @@ func statToAPI(n *node.Node) *api.NodeStat {
 	return &api.NodeStat{
 		Type:     n.Type().String(),
 		Size:     n.Size(),
-		Mode:     int32(n.Mode()),
-		Nlink:    int32(n.NLink()),
+		Mode:     n.Mode(),
+		Nlink:    n.NLink(),
 		Ino:      n.ID(),
 		UID:      apputils.OptString(n.UID()),
 		Gid:      apputils.OptString(n.GID()),
